Improved how the autodetect works for AMD and NVIDIA systems so that architecture specification is more robust.  Changed the data field for GPU memory to a float.

+++ hpc_launcher/systems/system.py
@@ -34,7 +34,7 @@ class SystemParams:
     # Vendor specific GPU compiler architecture
     gpu_arch: str
     # Number of GB of memory per GPU
-    mem_per_gpu: int
+    mem_per_gpu: float
     # # Physical number of CPUs per node
     # cpus_per_node: int
     # Number of NUMA domains
